refactor: New Project Example page container presenter 패턴으로 디렉토리 구조 변경 - New Project Example은 presenter component이므로 component 디렉토리 안으로 이동
