+++ src/pages/NewProjectExample/NewProjectExampleContainer.tsx
@@ -1,17 +1,14 @@
 import React from 'react';
 import { useLocation } from 'react-router-dom';
-import NewProjectExample from './NewProjectExample';
+import NewProjectExample from './componets/NewProjectExample';
 
 interface LocationState {
-  token:string;
-  platform:string;
+  token: string;
+  platform: string;
 }
-const NewProjectExampleContainer = () : JSX.Element => {
+const NewProjectExampleContainer = (): JSX.Element => {
   const location = useLocation<LocationState>();
-  const { token,platform } = location.state;
-  return (
-    <NewProjectExample {...{token,platform}} />
-  );
-}
+  const { token, platform } = location.state;
+  return <NewProjectExample {...{ token, platform }} />;
+};
 export default NewProjectExampleContainer;
-
